% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Suma.dos.numeros.r
\name{Suma.dos.numeros}
\alias{Suma.dos.numeros}
\title{Suma 2 Function}
\usage{
Suma.dos.numeros(a, b = 0)
}
\arguments{
\item{2}{numbers.}
}
\description{
This function allows you to add 2 numbers.
}
\examples{
Suma.dos.numeros()
}
\keyword{suma}

